common: add CopyImageTags attribute to CopyImage
When copying images from a region to another, originally we wouldn't
keep the tags in the destination image.

This is in contradiction with the documentation, since the docs point
out that the tags should be copied when copying an image between
regions.

This was not supported by the Amazon APIs, but is nowadays, therefore we
support it by default for this operation.